+++ reference/ir_reference_project.h
@@ -15,6 +15,7 @@ bool IR_ReferenceReadPersistentByte(uint32_t record_id, uint32_t offset, uint8_t
 void IR_ReferenceSetPersistFailureStep(uint32_t step);
 void IR_ReferenceEnablePersistPauseInjection(bool enable);
 void IR_ReferenceSimulateRecovery(void);
+void IR_ReferenceSimulateRestart(void);
 void IR_ReferenceResetPersistentStore(void);
 
 #endif /* IR_REFERENCE_PROJECT_H */